requirements.txt
Django>=1.9
django-modeltranslation==0.19.19
django-tinymce>=2.6.0
